+++ lib/src/web_utils_extensions.dart
@@ -2,11 +2,70 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:js_interop_utils/js_interop_utils.dart';
+import 'package:web/web.dart' as web;
 import 'package:web/web.dart';
+
+extension WebObjectExtension on Object? {
+  bool get isNode => asJSAny.isNode;
+
+  bool get isElement => asJSAny.isElement;
+
+  bool get isHTMLElement => asJSAny.isHTMLElement;
+}
+
+extension WebJSAnyExtension on JSAny? {
+  bool get isNode => isA<Node>();
+
+  bool get isElement => isA<Element>();
+
+  bool get isHTMLElement => isA<HTMLElement>();
+}
+
+extension JSObjectWebExtension on JSObject? {
+  bool get isNode => isA<Node>();
+
+  bool get isElement => isA<Element>();
+
+  bool get isHTMLElement => isA<HTMLElement>();
+}
+
+extension WebIterableExtension<T> on Iterable<T> {
+  Iterable<Element> whereElement() {
+    return whereJSAny().where((e) => e.isA<Element>()).cast<Element>();
+  }
+
+  Iterable<E> whereElementType<E extends Element>(Web<E> webType) {
+    return whereElement().where(webType.isOf).cast<E>();
+  }
+}
+
+extension WebIterableJSAnyExtension<T extends JSAny?> on Iterable<T> {
+  Iterable<Element> whereElement() {
+    return where((e) => e != null && e.isA<Element>()).cast<Element>();
+  }
+
+  Iterable<E> whereElementType<E extends Element>(Web<E> webType) {
+    return whereElement().whereElementType<E>(webType);
+  }
+}
+
+extension IterableElementExtension<T extends Element?> on Iterable<T> {
+  Iterable<Element> whereElement() {
+    return where((e) => e != null && e.isA<Element>()).cast<Element>();
+  }
+
+  Iterable<E> whereElementType<E extends Element>(Web<E> webType) {
+    return whereElement().where(webType.isOf).cast<E>();
+  }
+}
 
 extension IterableNodeExtension on Iterable<Node> {
   Iterable<Element> whereElement() =>
       where((e) => e.isA<Element>()).cast<Element>();
+
+  Iterable<E> whereElementType<E extends Element>(Web<E> webType) {
+    return whereElement().whereElementType<E>(webType);
+  }
 
   List<Element> toElements() => whereElement().toList();
 
@@ -17,14 +76,24 @@ extension IterableNodeExtension on Iterable<Node> {
 }
 
 extension DocumentExtension on Document {
-  T? select<T extends Element>(String selectors) {
+  T? select<T extends Element>(String selectors, [Web<T>? webType]) {
     var o = document.querySelector(selectors);
-    return o?.asElementOf<T>();
+
+    if (webType != null) {
+      return o?.asElementOf<T>(webType);
+    } else {
+      return o as T?;
+    }
   }
 
-  List<T> selectAll<T extends Element>(String selectors) {
+  List<T> selectAll<T extends Element>(String selectors, [Web<T>? webType]) {
     var l = document.querySelectorAll(selectors).whereElement();
-    return l.map((e) => e.asElementOf<T>()).nonNulls.toList();
+
+    if (webType != null) {
+      return l.map((e) => e.asElementOf<T>(webType)).nonNulls.toList();
+    } else {
+      return l.nonNulls.cast<T>().toList();
+    }
   }
 }
 
@@ -70,6 +139,31 @@ extension NodeExtension on Node {
     parentNode?.removeChild(this);
   }
 
+  void removeNodes(Iterable<Node> nodes) {
+    for (var node in nodes.toList(growable: false)) {
+      removeChild(node);
+    }
+  }
+
+  void removeNodeWhere(bool Function(Node node) test) {
+    final childNodes = this.childNodes;
+
+    for (var i = 0; i < childNodes.length;) {
+      var node = childNodes.item(i)!;
+      if (test(node)) {
+        final prevLength = childNodes.length;
+        if (node.isA<Element>()) {
+          (node as Element).remove();
+        } else {
+          removeChild(node);
+        }
+        assert(childNodes.length == prevLength - 1);
+      } else {
+        ++i;
+      }
+    }
+  }
+
   void clear() {
     if (isA<Element>()) {
       asElement.innerHTML = ''.toJS;
@@ -87,32 +181,142 @@ extension NodeExtension on Node {
   }
 }
 
-extension ElementNullableExtension on Element? {
-  T? asElementOf<T extends Element>() {
-    final self = this;
+enum Web<T> {
+  // ignore: constant_identifier_names
+  Element<web.Element>(_isElement, _castElement),
+  // ignore: constant_identifier_names
+  HTMLElement(_isHTMLElement, _castHTMLElement),
+  // ignore: constant_identifier_names
+  HTMLDivElement(_isHTMLDivElement, _castHTMLDivElement),
+  // ignore: constant_identifier_names
+  HTMLButtonElement(_isHTMLButtonElement, _castHTMLButtonElement),
+  // ignore: constant_identifier_names
+  HTMLTextAreaElement(_isHTMLTextAreaElement, _castHTMLTextAreaElement),
+  // ignore: constant_identifier_names
+  HTMLInputElement(_isHTMLInputElement, _castHTMLInputElement),
+  // ignore: constant_identifier_names
+  HTMLSelectElement(_isHTMLSelectElement, _castHTMLSelectElement),
+  // ignore: constant_identifier_names
+  HTMLAnchorElement(_isHTMLAnchorElement, _castHTMLAnchorElement),
+  // ignore: constant_identifier_names
+  HTMLImageElement(_isHTMLImageElement, _castHTMLImageElement),
+  // ignore: constant_identifier_names
+  HTMLSpanElement(_isHTMLSpanElement, _castHTMLSpanElement),
+  // ignore: constant_identifier_names
+  HTMLTableElement(_isHTMLTableElement, _castHTMLTableElement),
+  // ignore: constant_identifier_names
+  HTMLTableRowElement(_isHTMLTableRowElement, _castHTMLTableRowElement),
+  // ignore: constant_identifier_names
+  HTMLTableCellElement(_isHTMLTableCellElement, _castHTMLTableCellElement);
 
-    if (self == null || self.isUndefinedOrNull) {
-      return null;
-    } else if (T == HTMLElement) {
-      return self.isA<HTMLElement>() ? self as T : null;
-    } else if (T == HTMLDivElement) {
-      return self.isA<HTMLDivElement>() ? self as T : null;
-    } else if (T == HTMLSpanElement) {
-      return self.isA<HTMLSpanElement>() ? self as T : null;
-    } else if (T == HTMLInputElement) {
-      return self.isA<HTMLInputElement>() ? self as T : null;
-    } else if (T == HTMLLabelElement) {
-      return self.isA<HTMLLabelElement>() ? self as T : null;
-    } else if (T == HTMLBodyElement) {
-      return self.isA<HTMLBodyElement>() ? self as T : null;
-    } else {
-      try {
-        return self as T;
-      } catch (_) {
-        return null;
-      }
-    }
+  /////////////////
+
+  static bool _isElement(JSAny? o) => o.isA<web.Element>();
+
+  static web.Element? _castElement(JSAny? o) =>
+      o.isA<web.Element>() ? o as web.Element : null;
+
+  static bool _isHTMLElement(JSAny? o) => o.isA<web.HTMLElement>();
+
+  static web.HTMLElement? _castHTMLElement(JSAny? o) =>
+      o.isA<web.HTMLElement>() ? o as web.HTMLElement : null;
+
+  static bool _isHTMLDivElement(JSAny? o) => o.isA<web.HTMLDivElement>();
+
+  static web.HTMLDivElement? _castHTMLDivElement(JSAny? o) =>
+      o.isA<web.HTMLDivElement>() ? o as web.HTMLDivElement : null;
+
+  static bool _isHTMLButtonElement(JSAny? o) => o.isA<web.HTMLButtonElement>();
+
+  static web.HTMLButtonElement? _castHTMLButtonElement(JSAny? o) =>
+      o.isA<web.HTMLButtonElement>() ? o as web.HTMLButtonElement : null;
+
+  static bool _isHTMLTextAreaElement(JSAny? o) =>
+      o.isA<web.HTMLTextAreaElement>();
+
+  static web.HTMLTextAreaElement? _castHTMLTextAreaElement(JSAny? o) =>
+      o.isA<web.HTMLTextAreaElement>() ? o as web.HTMLTextAreaElement : null;
+
+  static bool _isHTMLInputElement(JSAny? o) => o.isA<web.HTMLInputElement>();
+
+  static web.HTMLInputElement? _castHTMLInputElement(JSAny? o) =>
+      o.isA<web.HTMLInputElement>() ? o as web.HTMLInputElement : null;
+
+  static bool _isHTMLSelectElement(JSAny? o) => o.isA<web.HTMLSelectElement>();
+
+  static web.HTMLSelectElement? _castHTMLSelectElement(JSAny? o) =>
+      o.isA<web.HTMLSelectElement>() ? o as web.HTMLSelectElement : null;
+
+  static bool _isHTMLAnchorElement(JSAny? o) => o.isA<web.HTMLAnchorElement>();
+
+  static web.HTMLAnchorElement? _castHTMLAnchorElement(JSAny? o) =>
+      o.isA<web.HTMLAnchorElement>() ? o as web.HTMLAnchorElement : null;
+
+  static bool _isHTMLImageElement(JSAny? o) => o.isA<web.HTMLImageElement>();
+
+  static web.HTMLImageElement? _castHTMLImageElement(JSAny? o) =>
+      o.isA<web.HTMLImageElement>() ? o as web.HTMLImageElement : null;
+
+  static bool _isHTMLSpanElement(JSAny? o) => o.isA<web.HTMLSpanElement>();
+
+  static web.HTMLSpanElement? _castHTMLSpanElement(JSAny? o) =>
+      o.isA<web.HTMLSpanElement>() ? o as web.HTMLSpanElement : null;
+
+  static bool _isHTMLTableElement(JSAny? o) => o.isA<web.HTMLTableElement>();
+
+  static web.HTMLTableElement? _castHTMLTableElement(JSAny? o) =>
+      o.isA<web.HTMLTableElement>() ? o as web.HTMLTableElement : null;
+
+  static bool _isHTMLTableRowElement(JSAny? o) =>
+      o.isA<web.HTMLTableRowElement>();
+
+  static web.HTMLTableRowElement? _castHTMLTableRowElement(JSAny? o) =>
+      o.isA<web.HTMLTableRowElement>() ? o as web.HTMLTableRowElement : null;
+
+  static bool _isHTMLTableCellElement(JSAny? o) =>
+      o.isA<web.HTMLTableCellElement>();
+
+  static web.HTMLTableCellElement? _castHTMLTableCellElement(JSAny? o) =>
+      o.isA<web.HTMLTableCellElement>() ? o as web.HTMLTableCellElement : null;
+
+  ///////////////////////
+
+  final bool Function(JSAny? o) _isA;
+
+  final T? Function(JSAny? o) _cast;
+
+  const Web(this._isA, this._cast);
+
+  bool isOf(Object? o) {
+    var jsObj = o.asJSObject;
+    if (jsObj == null) return false;
+    return _isA(jsObj);
   }
+
+  T? castNullable(Object? o) {
+    var jsObj = o.asJSObject;
+    if (jsObj == null) return null;
+    return _cast(jsObj);
+  }
+
+  T cast(Object? o) {
+    var jsObj = o.asJSObject;
+    if (jsObj == null) throw ArgumentError("Can't cast `null` to type `$name`");
+    return _cast(jsObj) ??
+        (throw ArgumentError("Can't cast `$o` to type `$name`"));
+  }
+
+  @override
+  String toString() => '`web` type `$name`';
+}
+
+extension ElementNullableExtension on Element? {
+  HTMLElement? get asHTMLElement =>
+      isA<HTMLElement>() ? this as HTMLElement : null;
+
+  bool isElementOf<T extends Element>(Web<T> webType) => webType.isOf(this);
+
+  T? asElementOf<T extends Element>(Web<T> webType) => webType.cast(this);
 
   bool get isCanvasImageSource {
     final node = this;
@@ -128,8 +332,25 @@ extension ElementExtension on Element {
 
   set text(String? text) => textContent = text;
 
-  HTMLElement? get asHTMLElement =>
+  HTMLElement get asHTMLElement => this as HTMLElement;
+
+  HTMLElement? get asHTMLElementChecked =>
       isA<HTMLElement>() ? this as HTMLElement : null;
+
+  /// Alias to [document.querySelector] casting to [T].
+  T? querySelectorTyped<T extends Element>(String selectors, Web<T> webType) {
+    var self = this;
+    var elem = self.querySelector(selectors);
+    return elem.isElementOf<T>(webType) ? elem as T? : null;
+  }
+
+  /// Alias to [querySelectorAll] casting elements to [T].
+  List<T> querySelectorAllTyped<T extends Element>(
+      String selectors, Web<T> webType) {
+    var self = this;
+    var list = self.querySelectorAll(selectors);
+    return list.whereElementType<T>(webType).toList();
+  }
 
   Element insertChild(int index, Element node) {
     final children = this.children;
@@ -160,13 +381,110 @@ extension ElementExtension on Element {
   void clear() {
     innerHTML = ''.toJS;
   }
+
+  CSSStyleDeclaration getComputedStyle() => window.getComputedStyle(this);
+
+  void click() => asHTMLElementChecked?.click();
+
+  bool focus() {
+    final htmlElement = asHTMLElementChecked;
+    if (htmlElement != null) {
+      htmlElement.focus();
+      return true;
+    }
+    return false;
+  }
+
+  CSSStyleDeclaration? get style => asHTMLElementChecked?.style;
+
+  bool get hidden => asHTMLElementChecked?.hidden?.dartify() == true;
+
+  set hidden(bool? value) => asHTMLElementChecked?.hidden = value?.toJS;
+
+  DOMTokenList? get classList => asHTMLElementChecked?.classList;
+
+  bool dispatchEventOfType(
+    String type, {
+    bool? bubbles,
+    bool? cancelable,
+    bool? composed,
+  }) {
+    var eventInit = bubbles != null || cancelable != null || composed != null
+        ? EventInit(
+            bubbles: bubbles ?? false,
+            cancelable: cancelable ?? false,
+            composed: composed ?? false,
+          )
+        : null;
+
+    var event = eventInit != null ? Event(type, eventInit) : Event(type);
+    return dispatchEvent(event);
+  }
+
+  bool dispatchChangeEvent() => dispatchEventOfType('change');
+
+  void setAttributes(Map<String, String> attrs) {
+    for (var e in attrs.entries) {
+      setAttribute(e.key, e.value);
+    }
+  }
+}
+
+extension HTMLElementExtension on HTMLElement {
+  static final DOMParser _domParser = DOMParser();
+
+  List<Node> appendHTML(String html) {
+    var doc = _domParser.parseFromString(html.toJS, 'text/html');
+    var root = doc.body ?? doc.documentElement;
+    var nodes = root?.childNodes.toList() ?? [];
+    appendNodes(nodes);
+    return nodes;
+  }
+
+  bool get isHidden {
+    var v = hidden;
+    if (v == null) return false;
+
+    if (v.isUndefinedOrNull) {
+      return false;
+    } else if (v.isA<JSBoolean>()) {
+      return (v as JSBoolean).toDart;
+    } else if (v.isA<JSString>()) {
+      var s = (v as JSString).toDart.toLowerCase().trim();
+      return s == '' || s == 'hidden';
+    }
+
+    return false;
+  }
+}
+
+extension HTMLSelectElementExtension on HTMLSelectElement {
+  /// Selects an [index] and triggers the `change` event.
+  /// See [dispatchChangeEvent] and [selectedIndex].
+  bool selectIndex(int index) {
+    selectedIndex = index;
+    return dispatchChangeEvent();
+  }
+
+  List<HTMLOptionElement> get selectedOptionsSafe {
+    try {
+      return selectedOptions
+          .toList()
+          .where((e) => e.isA<HTMLOptionElement>())
+          .cast<HTMLOptionElement>()
+          .toList();
+    } catch (_) {
+      return [];
+    }
+  }
 }
 
 extension HTMLOptionsCollectionExtension on HTMLOptionsCollection {
   Iterable<HTMLOptionElement> toIterable() =>
-      List.generate(length, (i) => item(i)).whereType<HTMLOptionElement>();
+      Iterable.generate(length, (i) => item(i) as HTMLOptionElement);
 
-  List<HTMLOptionElement> toList() => toIterable().toList();
+  List<HTMLOptionElement> toList() =>
+      List.generate(length, (i) => item(i) as HTMLOptionElement);
 
   bool get isEmpty => length == 0;
 
@@ -174,10 +492,9 @@ extension HTMLOptionsCollectionExtension on HTMLOptionsCollection {
 }
 
 extension TouchListExtension on TouchList {
-  Iterable<Touch> toIterable() =>
-      List.generate(length, (i) => item(i)).whereType<Touch>();
+  Iterable<Touch> toIterable() => Iterable.generate(length, (i) => item(i)!);
 
-  List<Touch> toList() => toIterable().toList();
+  List<Touch> toList() => List.generate(length, (i) => item(i)!);
 
   bool get isEmpty => length == 0;
 
@@ -185,10 +502,9 @@ extension TouchListExtension on TouchList {
 }
 
 extension FileListExtension on FileList {
-  Iterable<File> toIterable() =>
-      List.generate(length, (i) => item(i)).whereType<File>();
+  Iterable<File> toIterable() => Iterable.generate(length, (i) => item(i)!);
 
-  List<File> toList() => toIterable().toList();
+  List<File> toList() => List.generate(length, (i) => item(i)!);
 
   bool get isEmpty => length == 0;
 
@@ -202,10 +518,9 @@ extension CSSStyleDeclarationExtension on CSSStyleDeclaration {
 }
 
 extension CSSRuleListExtension on CSSRuleList {
-  Iterable<CSSRule> toIterable() =>
-      List.generate(length, (i) => item(i)).whereType<CSSRule>();
+  Iterable<CSSRule> toIterable() => Iterable.generate(length, (i) => item(i)!);
 
-  List<CSSRule> toList() => toIterable().toList();
+  List<CSSRule> toList() => List.generate(length, (i) => item(i)!);
 
   bool get isEmpty => length == 0;
 
@@ -224,9 +539,9 @@ extension CanvasRenderingContext2DExtension on CanvasRenderingContext2D {
 
 extension DOMTokenListExtension on DOMTokenList {
   Iterable<String> toIterable() =>
-      List.generate(length, (i) => item(i)?.toDartFix).nonNulls;
+      Iterable.generate(length, (i) => item(i)!.toDartFix);
 
-  List<String> toList() => toIterable().toList();
+  List<String> toList() => List.generate(length, (i) => item(i)!.toDartFix);
 
   bool get isEmpty => length == 0;
 
@@ -262,9 +577,9 @@ extension DOMTokenListExtension on DOMTokenList {
 }
 
 extension NamedNodeMapExtension on NamedNodeMap {
-  Iterable<Attr> toIterable() => List.generate(length, (i) => item(i)).nonNulls;
+  Iterable<Attr> toIterable() => Iterable.generate(length, (i) => item(i)!);
 
-  List<Attr> toList() => toIterable().toList();
+  List<Attr> toList() => List.generate(length, (i) => item(i)!);
 
   Iterable<T> whereType<T>() => toIterable().whereType<T>();
 
@@ -281,10 +596,9 @@ extension NamedNodeMapExtension on NamedNodeMap {
 }
 
 extension HTMLCollectionExtension on HTMLCollection {
-  Iterable<Element> toIterable() =>
-      List.generate(length, (i) => item(i)).nonNulls;
+  Iterable<Element> toIterable() => Iterable.generate(length, (i) => item(i)!);
 
-  List<Element> toList() => toIterable().toList();
+  List<Element> toList() => List.generate(length, (i) => item(i)!);
 
   Iterable<T> whereType<T>() => toIterable().whereType<T>();
 
@@ -305,11 +619,15 @@ extension HTMLCollectionExtension on HTMLCollection {
 }
 
 extension NodeListExtension on NodeList {
-  Iterable<Node> toIterable() => List.generate(length, (i) => item(i)).nonNulls;
+  Iterable<Node> toIterable() => Iterable.generate(length, (i) => item(i)!);
 
-  List<Node> toList() => toIterable().toList();
+  List<Node> toList() => List.generate(length, (i) => item(i)!);
 
   Iterable<T> whereType<T extends Node>() => toIterable().whereType<T>();
+
+  Iterable<E> whereElementType<E extends Element>(Web<E> webType) {
+    return whereElement().where((e) => e.isElementOf<E>(webType)).cast<E>();
+  }
 
   Iterable<Element> whereElement() => whereType<Element>();
 
@@ -335,13 +653,25 @@ extension NodeListExtension on NodeList {
   }
 }
 
+extension HTMLTableElementExtension on HTMLTableElement {
+  HTMLTableRowElement appendRow() => insertRow(-1);
+}
+
+extension HTMLTableSectionElementExtension on HTMLTableSectionElement {
+  HTMLTableRowElement appendRow() => insertRow(-1);
+}
+
+extension HTMLTableRowElementExtension on HTMLTableRowElement {
+  HTMLTableCellElement appendCell() => insertCell(-1);
+}
+
 extension StorageExtension on Storage {
   bool get isEmpty => length == 0;
 
   bool get isNotEmpty => !isEmpty;
 
   Iterable<String> get keys =>
-      List.generate(length, (i) => key(i)?.toDartFix).nonNulls;
+      Iterable.generate(length, (i) => key(i)?.toDartFix).nonNulls;
 
   bool remove(String key) {
     var prevLng = length;
@@ -448,6 +778,63 @@ extension MouseEventExtension on MouseEvent {
   Point get pagePoint => Point(pageX, pageY);
 
   Point get screenPoint => Point(screenX, screenY);
+}
+
+extension TouchExtension on Touch {
+  Point get clientPoint => Point(clientX, clientY);
+
+  Point get radiusPoint => Point(radiusX, radiusY);
+
+  Point get pagePoint => Point(pageX, pageY);
+
+  Point get screenPoint => Point(screenX, screenY);
+}
+
+extension FileReaderExtension on FileReader {
+  Stream<ProgressEvent> get onLoad =>
+      EventStreamProviders.loadEvent.forTarget(this);
+
+  Stream<ProgressEvent> get onLoadStart =>
+      EventStreamProviders.loadStartEvent.forTarget(this);
+
+  Stream<ProgressEvent> get onError =>
+      EventStreamProviders.errorEvent.forTarget(this);
+
+  Stream<ProgressEvent> get onAbort =>
+      EventStreamProviders.abortEvent.forTarget(this);
+}
+
+extension WorkerGlobalScopeExtension on WorkerGlobalScope {
+  web.$Console get console => web.console;
+}
+
+extension WindowExtension on Window {
+  web.$Console get console => web.console;
+
+  Stream<Event> get onResize =>
+      EventStreamProviders.resizeEvent.forTarget(this);
+
+  Stream<Event> get onHashChange =>
+      EventStreamProviders.hashChangeEvent.forTarget(this);
+
+  Stream<DeviceOrientationEvent> get onDeviceOrientation =>
+      EventStreamProviders.deviceOrientationEvent.forTarget(this);
+
+  Stream<TouchEvent> get onTouchStart =>
+      EventStreamProviders.touchStartEvent.forTarget(this);
+
+  Stream<TouchEvent> get onTouchEndEvent =>
+      EventStreamProviders.touchEndEvent.forTarget(this);
+
+  Stream<TouchEvent> get onTouchEnterEvent =>
+      EventStreamProviders.touchEnterEvent.forTarget(this);
+
+  Stream<TouchEvent> get onTouchLeaveEvent =>
+      EventStreamProviders.touchLeaveEvent.forTarget(this);
+}
+
+extension DOMRectExtension on DOMRect {
+  Rectangle<double> toRectangle() => Rectangle<double>(x, y, width, height);
 }
 
 extension _CompleterExtension<T> on Completer<T> {